rename module add initial whisper support
